Fix windows build flags
We can't specify more than one /arch: flag, or at least the one we actually want needs to be first or last.

This also uses AVX2 for fma and f16c, which appears to be necessary. This is sketchy, because we don't want other parts of AVX2 in some cases, but AFAICT this is the best we can do with MSVC's options. We just have to hope that the compiler doesn't automatically use some more advanced instructions than we actually want.

PiperOrigin-RevId: 843401782