Add `fma`, `concat`, and more `convert` functions to SIMD headers
To facilitate this, the headers are split into internal and external headers, where the external headers might implement the same operation in two different ways. For example, converting f16x16 to f32x16 is implemented internally as one avx512 operation and vector, but two avx operations and vectors. This allows calling code to be written the same for avx512 and avx, while generating good code for both.

This change also uses the new SIMD functionality to clean up the reduce kernels. In particular, we now call `convert` in generic `reduce_add` implementation. This is often a no-op, but when it is not it eliminates the need for many `reduce_add` overloads.

There is one change in performance from this change:
```
bench/sum_squared_bf16_fp32_sse2/real_time       [256x1x256x1] 5.354µ ± 0%   6.745µ ± 1%  +26.00% (p=0.002 n=6)
```

I looked at this for quite a while, there's barely any difference in what we generate, this is just tickling the compiler in a way that it doesn't like. Since we are going to change how the non-contiguous reductions are done shortly, I think we should ignore it.

PiperOrigin-RevId: 844771662