Pass down the original kernel and bias buffers when creating a fully connected operation.
Some fully connected create functions convert buffers on the fly (for instance
`xnn_create_fully_connected_nc_f32_f16`). This breaks buffer identification for
the delegate (as it relies on pointer values). In order to avoid this, we pass
down the original buffer pointers and use these to create the cache key.

PiperOrigin-RevId: 845185501