Rewrite SIMD tests to make a new test for each architecture
Currently, these tests are written to skip tests if the test requires an architecture not currently supported. This is problematic in at least two ways:
- There is code that runs before we reach our runtime feature flag check (e.g. gtest code) that is compiled with the feature flag, that may not be supported.
- Code that is inlined into these translation units may get deduped by the linker, and the instance that is linked might use a feature flag that is not supported by the current machine.

This works around the problem by making a separate test for each target, and then skipping the test suite entirely if it is unsupported.

There is still a risk that code in/before main is compiled using unavailable instructions, but that is a much lower risk.

PiperOrigin-RevId: 845538611