Enable transpose_a kernels for num_k_dims > 1
This change makes `transpose_a` able to transpose dimension 0 and any other dimension, instead of only dimension 0 and 1.

This enables using transpose_a kernels for convolutions. This change by itself only enables us to transpose A, but *after* the `stencil_copy` op. Ideally, we would reorder the `transpose_a` and `stencil_copy` ops in the graph, so we don't "realize" the im2col into memory.

PiperOrigin-RevId: 846018481