Include bias in initializer of dot for f32.
This fixes a bug where the bias was outside of the dot followed by a convert from f32 to f32.

For the f32 case, it makes sense to initialize the dot accumulator with the bias.

For the quantized case, we want to keep the bias add outside of the dot in order to enable fusions and have the bias use the right scale.

Due to this divergence, this change also refactors `define_xnn_dot` to use two branches - one for f32, one for quantized.

PiperOrigin-RevId: 847858417