Fix `tile_k > 1` case for rewriting `transpose_a(stencil_copy(x))` -> `stencil_copy(transpose_a(x))`
This implementation of transpose_a adjusts the elem_size of the output, so we don't need to adjust the dilation to implement the stencil.

This also disables storage folding of the transposed dimensions, which avoids a spurious assert, and is necessary anyways (we can't handle folded dimensions in the transpose operation).

PiperOrigin-RevId: 849222018